# oas/OAS.yaml
openapi: 3.1.0
info:
  title: Multi-Tenant Salon API
  description: |
    REST API for the multi-tenant SaaS salon/salon management platform.

    ## Module / Access Hierarchy

    The API is split into four distinct path namespaces:

    | Namespace | Base Path | Who uses it |
    |---|---|---|
    | **Salon Onboarding** | `/api/salon-onboarding` | Anyone registering a new salon tenant; platform-level listing |
    | **Customer** | `/api/salon/{salonId}/...` | Public/customer-facing — browse, book, read |
    | **Admin** | `/api/salon-admin/{salonId}/...` | Salon owner/administrator — manage, configure, operate. Also callable by a platform super-admin for any salon, since super-admin management of a specific salon reuses these endpoints rather than duplicating them. |
    | **Super Admin** | `/api/salon-super-admin/...` | Platform super-admin — cross-tenant management of all salons |
    | **Staff Portal** | `/api/salon-staff/...` | Authenticated staff member — self-service profile, appointments, personal holidays |
    | **Utility** | `/api/salon-utility/...` | Any consumer needing reference data (countries with embedded currency info) |

    ### Customer sub-paths
    - `/api/salon/{salonId}/services/...` — browse the service catalog
    - `/api/salon/{salonId}/staff/...` — view staff members
    - `/api/salon/{salonId}/booking/...` — slot discovery and appointment creation
    - `/api/salon/{salonId}/website` — public website theme
...

    ### Admin sub-paths
...

...
servers:
...
paths:
...
  /api/salon/{salonIdOrHandler}:
    get:
      tags: [Salon]
      summary: Get salon by ID or handler (public)
      description: |
        Accepts either a UUID (`a1b2c3d4-e5f6-7890-abcd-ef1234567890`) or a handler slug
        (`glam-salon`). UUID lookup is tried first; if the value is not a valid UUID the
        request falls back to a handler lookup.
      operationId: getSalon
      parameters:
        - name: salonIdOrHandler
          in: path
          required: true
          schema:
            type: string
          description: Salon UUID or URL-friendly handler slug
      responses:
        "200":
          description: Salon details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Salon"
        "404":
          $ref: "#/components/responses/NotFound"

  # ─── Customer — Services ──────────────────────────────────────────────────────
  /api/salon/{salonId}/services:
    parameters:
      - $ref: "#/components/parameters/SalonId"

    get:
      tags: [Services]
      summary: List services for a salon (public)
      operationId: listServices
      responses:
        "200":
          description: List of active services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServiceItem"
  /api/salon/{salonId}/services/{serviceId}:
    parameters:
      - $ref: "#/components/parameters/SalonId"
      - $ref: "#/components/parameters/ServiceId"

    get:
      tags: [Services]
      summary: Get a service by ID (public)
      operationId: getService
      responses:
        "200":
          description: Service details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceItem"
        "404":
          $ref: "#/components/responses/NotFound"

  # ─── Customer — Staff ─────────────────────────────────────────────────────────
  /api/salon/{salonId}/staff:
    parameters:
      - $ref: "#/components/parameters/SalonId"

    get:
      tags: [Staff]
      summary: List staff members for a salon (public)
      operationId: listStaff
      responses:
        "200":
          description: List of staff members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StaffMember"
  /api/salon/{salonId}/staff/{staffId}:
    parameters:
      - $ref: "#/components/parameters/SalonId"
      - $ref: "#/components/parameters/StaffId"

    get:
      tags: [Staff]
      summary: Get a staff member by ID (public)
      operationId: getStaffMember
      responses:
        "200":
          description: Staff member profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"
        "404":
          $ref: "#/components/responses/NotFound"

  # ─── Customer — Booking ───────────────────────────────────────────────────────
...
  /api/salon-admin/{salonId}/services:
    parameters:
      - $ref: "#/components/parameters/SalonId"

    get:
      tags: [AdminServices]
      summary: List services (admin)
      operationId: adminListServices
      responses:
        "200":
          description: List of service items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServiceItem"

    post:
      tags: [AdminServices]
      summary: Add a service to the catalog (admin)
      operationId: addService
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddServiceRequest"
      responses:
        "201":
          description: Service created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceItem"
  /api/salon-admin/{salonId}/services/{serviceId}:
    parameters:
      - $ref: "#/components/parameters/SalonId"
      - $ref: "#/components/parameters/ServiceId"

    get:
      tags: [AdminServices]
      summary: Get a service by ID (admin)
      operationId: adminGetService
      responses:
        "200":
          description: Service details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceItem"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      tags: [AdminServices]
      summary: Update a service (admin)
      description: Admin-only. Updates pricing, duration, category, assigned staff, and active status.
      operationId: updateService
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateServiceRequest"
      responses:
        "200":
          description: Service updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceItem"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      tags: [AdminServices]
      summary: Delete a service (admin)
      operationId: deleteService
      responses:
        "204":
          description: Service deleted
        "404":
          $ref: "#/components/responses/NotFound"

  # ─── Admin — Staff ────────────────────────────────────────────────────────────
  /api/salon-admin/{salonId}/staff:
    parameters:
      - $ref: "#/components/parameters/SalonId"

    get:
      tags: [AdminStaff]
      summary: List staff members (admin)
      operationId: adminListStaff
      responses:
        "200":
          description: List of staff members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StaffMember"

    post:
      tags: [AdminStaff]
      summary: Onboard a new staff member (admin)
      operationId: onboardStaff
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnboardStaffRequest"
      responses:
        "201":
          description: Staff member created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"
  /api/salon-admin/{salonId}/staff/{staffId}:
    parameters:
      - $ref: "#/components/parameters/SalonId"
      - $ref: "#/components/parameters/StaffId"

    get:
      tags: [AdminStaff]
      summary: Get a staff member by ID (admin)
      operationId: adminGetStaffMember
      responses:
        "200":
          description: Staff member details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      tags: [AdminStaff]
      summary: Update a staff member (admin)
      description: Admin-only. Updates name, contact, role, status, specializations, photos, bio, and booking availability.
      operationId: updateStaffMember
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStaffRequest"
      responses:
        "200":
          description: Staff member updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      tags: [AdminStaff]
      summary: Remove a staff member (admin)
      description: Admin-only. Removes a staff member from the roster. Existing bookings for this member are cascade-deleted. Owner-staff entries (`isOwner = true`) cannot be deleted.
      operationId: deleteStaffMember
      responses:
        "204":
          description: Staff member removed
        "404":
          $ref: "#/components/responses/NotFound"

  # ─── Admin — Staff Availability ───────────────────────────────────────────────
...
  /api/salon-super-admin/salons/{id}:
    get:
      tags: [SuperAdmin]
      summary: Get a salon by ID (super-admin)
      operationId: superAdminGetSalon
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: |
            Salon UUID or handler slug — same resolution rule as the `salonId` parameter used
            elsewhere (UUID tried first, falls back to a handler lookup; 404 if neither matches).
      responses:
        "200":
          description: Salon found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Salon"
        "404":
          description: Salon not found

    delete:
      tags: [SuperAdmin]
      summary: Disable a salon (super-admin)
      description: Soft-deletes the salon (sets status to DISABLED). Data is preserved and the salon can be re-enabled.
      operationId: superAdminDisableSalon
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: |
            Salon UUID or handler slug — same resolution rule as the `salonId` parameter used
            elsewhere (UUID tried first, falls back to a handler lookup; 404 if neither matches).
      responses:
        "200":
          description: Salon disabled; returns updated salon
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Salon"
        "404":
          description: Salon not found
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT issued by the identity provider after the admin authenticates.
        **Not yet enforced** — authentication/authorisation is a planned future addition.
        Currently the `salonId` path parameter is the only enforcement in place.
  parameters:
    SalonId:
      name: salonId
      in: path
      required: true
      schema:
        type: string
      description: |
        Salon UUID (`a1b2c3d4-e5f6-7890-abcd-ef1234567890`) or handler slug (`glam-salon`).
        UUID form is tried first; if the value isn't a valid UUID it falls back to a handler
        lookup — the same rule `GET /api/salon/{salonIdOrHandler}` uses. Returns `404` if
        neither matches.
    ServiceId:
      name: serviceId
      in: path
      required: true
      schema:
        type: integer
        format: int64
      description: Service item ID
    StaffId:
      name: staffId
      in: path
      required: true
      schema:
        type: integer
        format: int64
      description: Staff member ID
...
  responses:
    NotFound:
      description: Resource not found
  schemas:
...
    Salon:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        handler:
          type: string
          description: URL-friendly slug
        owner:
          $ref: "#/components/schemas/Owner"
        location:
          $ref: "#/components/schemas/Location"
        contact:
          $ref: "#/components/schemas/ContactInfo"
        operatingHours:
          type: array
          items:
            $ref: "#/components/schemas/OperatingHours"
        features:
          type: array
          items:
            $ref: "#/components/schemas/SalonFeature"
          description: Serialized as a flat array of feature strings
          example: ["BOOKING", "ANALYTICS"]
        bookingAdvanceDays:
          type: integer
          description: How many days in advance customers can book. Defaults to 60.
          default: 60
        businessRegistrationId:
          type: string
          description: Company/business registration number (e.g. CVR, EIN, Company Number).
        showBusinessId:
          type: boolean
          description: Whether to display the registration number on the public website. Defaults to false.
          default: false
        termsAccepted:
          type: boolean
          description: Whether the salon owner has accepted the terms and conditions. Defaults to false.
          default: false
        termsAcceptedAt:
          type: string
          format: date-time
          description: Timestamp when the terms were accepted. Set automatically on creation when termsAccepted is true.
          readOnly: true
        bookingRequiresConfirmation:
          type: boolean
          description: >
            When true, new bookings are created with status PENDING and must be manually confirmed
            by an admin. When false (default), bookings are auto-confirmed immediately (status CONFIRMED).
          default: false
        businessIdLabel:
          type: string
          description: >
            Country-specific label for the registration number (e.g. "CVR Number" for Denmark,
            "EIN" for United States). Derived automatically from location.country on create/update;
            read-only in responses.
          readOnly: true
        createdAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [ACTIVE, DISABLED]
          description: Salon lifecycle status. Defaults to ACTIVE on creation.
...
    Owner:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
    Location:
      type: object
      properties:
        address:
          type: string
        city:
          type: string
        state:
          type: string
        country:
          type: string
        zipCode:
          type: string
    ContactInfo:
      type: object
      properties:
        phone:
          type: string
        email:
          type: string
          format: email
        website:
          type: string
          format: uri
    OperatingHours:
      type: object
      properties:
        day:
          type: string
          enum: [MONDAY, TUESDAY, WEDNESDAY, THURSDAY, FRIDAY, SATURDAY, SUNDAY]
        openTime:
          type: string
          description: Opening time in HH:mm format
          example: "09:00"
        closeTime:
          type: string
          description: Closing time in HH:mm format
          example: "18:00"
        closed:
          type: boolean
          default: false
    SalonFeature:
      type: string
      enum:
        - STATIC_WEBSITE
        - BOOKING
        - MEMBERSHIP
        - WEBSHOP
        - ANALYTICS
        - LOYALTY_PROGRAM

    # ── Service ───────────────────────────────────────────────────────────────
    ServiceItem:
      type: object
      properties:
        id:
          type: integer
          format: int64
        salonId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        price:
          type: ["number", "null"]
          description: null for pay-as-you-go services
        currency:
          type: ["string", "null"]
          example: USD
          description: null when price is null
        durationMinutes:
          type: ["integer", "null"]
          description: null for pay-as-you-go services; UI defaults to 30 min
        category:
          $ref: "#/components/schemas/ServiceCategory"
        active:
          type: boolean
        assignedStaffIds:
          type: array
          items:
            type: string
          description: Serialized as a flat array of staff ID strings
        createdAt:
          type: string
          format: date-time
    AddServiceRequest:
      type: object
      required: [name, category]
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: ["number", "null"]
          description: Omit or send null for a pay-as-you-go service
        currency:
          type: ["string", "null"]
          description: ISO 4217; omit or send null when price is null
        durationMinutes:
          type: ["integer", "null"]
          description: Omit or send null for pay-as-you-go; slot engine defaults to 30 min
        category:
          $ref: "#/components/schemas/ServiceCategory"
        assignedStaffIds:
          type: array
          items:
            type: string
    UpdateServiceRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: ["number", "null"]
          description: null for pay-as-you-go services
        currency:
          type: ["string", "null"]
          description: null when price is null
        durationMinutes:
          type: ["integer", "null"]
          description: null for pay-as-you-go; UI defaults to 30 min
        category:
          $ref: "#/components/schemas/ServiceCategory"
        active:
          type: boolean
        assignedStaffIds:
          type: array
          items:
            type: string
    ServiceCategory:
      type: string
      enum:
        - HAIR
        - MAKEUP
        - NAILS
        - SKIN_CARE
        - BEARD
        - MASSAGE
        - WAXING
        - OTHER

    # ── Staff ─────────────────────────────────────────────────────────────────
    StaffMember:
      type: object
      properties:
        id:
          type: integer
          format: int64
        salonId:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        role:
          $ref: "#/components/schemas/StaffRole"
        status:
          $ref: "#/components/schemas/StaffStatus"
        isOwner:
          type: boolean
          description: >
            True when this staff member was auto-created from the salon owner's profile.
            Owner-staff entries cannot be deleted; the owner may opt out of the booking
            calendar by setting availableForBooking to false.
          default: false
        availableForBooking:
          type: boolean
          description: >
            When false, this staff member is excluded from slot discovery and will not
            be offered to customers during booking.
          default: true
        specializations:
          type: array
          items:
            type: string
          description: Serialized as a flat array of specialization strings
        photoUrls:
          type: array
          items:
            type: string
            format: uri
          description: Photo URLs shown on the public website
        bio:
          type: string
          description: Short bio shown on the public website
        createdAt:
          type: string
          format: date-time
...
    OnboardStaffRequest:
      type: object
      required: [name, role]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        role:
          $ref: "#/components/schemas/StaffRole"
        specializations:
          type: array
          items:
            type: string
        photoUrls:
          type: array
          items:
            type: string
            format: uri
        bio:
          type: string
    UpdateStaffRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        role:
          $ref: "#/components/schemas/StaffRole"
        status:
          $ref: "#/components/schemas/StaffStatus"
        availableForBooking:
          type: boolean
          description: Set to false to remove this staff member from slot discovery
          default: true
        specializations:
          type: array
          items:
            type: string
        photoUrls:
          type: array
          items:
            type: string
            format: uri
        bio:
          type: string
    StaffRole:
      type: string
      enum:
        - MANAGER
        - STYLIST
        - COLORIST
        - MAKEUP_ARTIST
        - NAIL_TECHNICIAN
        - RECEPTIONIST
        - ASSISTANT
    StaffStatus:
      type: string
      enum:
        - ACTIVE
        - INACTIVE
        - ON_LEAVE

    # ── Availability ─────────────────────────────────────────────────────────
...